Decompose some methods
First, remove as much logic as possible from the curl's `write_function`
callback and extract the code for getting the current standards file.

In 'update_config' the logic is much the same, except I'm doing it
outside the curl function, so I'm working with the file instead of the
data stream.