Refactor (#12)
* Decompose some methods

First, remove as much logic as possible from the curl's `write_function`
callback and extract the code for getting the current standards file.

In 'update_config' the logic is much the same, except I'm doing it
outside the curl function, so I'm working with the file instead of the
data stream.

* Split out new file for cli

We're going to be adding behavior to this, so I want to get it out of
main

* Update clap

* Extract options

To start with I'm building out the default options, beginning with the
target filename and source url.

The options are not in use yet.

Note: I have no idea how to decompose Rust code into separate files.
`extern crate`, `use`, `mod`. No idea what's going on here. I'm mostly
just trying stuff out until it compiles.

* Cleanup (?) inclusions

Still not quite sure how this stuff should work. Here I'm mostly copying
the example of Lightning

https://github.com/chriskrycho/lightning-rs

So if he's doing it wrong then so am I.

Generally I'm trying to move all of the crate/mod stuff to `src/lib.rs`.
The main function still needs to declare the cli module and use the cli
function.

* Provide options to `run`

Slowly, ever so slowly, getting to a point where we can use options
inside of the run function.

* Use source_url option to get reek standards

* Use target_filepath option in update_config

There's still too much going on in update_config, but at least the
filename is no longer hardcoded in there.